+++ alitheia/core/src/main/java/eu/sqooss/impl/service/webadmin/ProjectDeleteJob.java
@@ -57,7 +57,7 @@ public class ProjectDeleteJob extends Job {
     }
 
     @Override
-    public int priority() {
+    public long priority() {
         return 0xff;
     }
