+++ plug-ins/bugzilla/src/main/java/eu/sqooss/plugins/bugzilla/BugzillaXMLJob.java
@@ -72,7 +72,7 @@ public class BugzillaXMLJob extends Job {
     }
     
     @Override
-    public int priority() {
+    public long priority() {
         return 3;
     }
